// packages/core/src/types/dataGridTypes.d.ts
export interface ISelectionRange {
    startRow: number;
    startCol: number;
    endRow: number;
    endCol: number;
}